requirements.txt
alabaster==0.7.12
arrow==0.15.4
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
codecov==2.0.15
construct==2.9.45
coverage==5.0
cycler==0.10.0
docutils==0.15.2
Flask==1.1.1
furl==2.1.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
more-itertools==7.2.0
numpy==1.18.2
orderedmultidict==1.0.1
packaging==19.2
pika==1.1.0
Pillow==7.0.0
pluggy==0.13.0
protobuf==3.11.3
psycopg2-binary==2.8.4
py==1.8.0
Pygments==2.5.2
pyparsing==2.4.2
pytest==5.2.2
pytest-cov==2.8.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.3.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
